+++ test/internal/Utils.t.sol
@@ -2,9 +2,10 @@
 pragma solidity ^0.8.20;
 
 import {Test} from "forge-std/Test.sol";
-import {strings} from "solidity-stringutils/src/strings.sol";
 
 import {Utils, ContractInfo} from "openzeppelin-foundry-upgrades/internal/Utils.sol";
+
+import {StringFinder} from "openzeppelin-foundry-upgrades/internal/StringFinder.sol";
 
 import {MyContractName} from "../contracts/MyContractFile.sol";
 
@@ -103,7 +104,7 @@ contract UtilsTest is Test {
         assertEq(Utils.getOutDir(), "out");
     }
 
-    using strings for *;
+    using StringFinder for string;
 
     function testGetBuildInfoFile() public {
         ContractInfo memory contractInfo = Utils.getContractInfo("Greeter.sol", "out");
@@ -113,8 +114,8 @@ contract UtilsTest is Test {
             "out"
         );
 
-        assertTrue(buildInfoFile.toSlice().startsWith("out/build-info".toSlice()));
-        assertTrue(buildInfoFile.toSlice().endsWith(".json".toSlice()));
+        assertTrue(buildInfoFile.startsWith("out/build-info"));
+        assertTrue(buildInfoFile.endsWith(".json"));
     }
 
     function testToBashCommand() public pure {